+++ packages/commitlint-config/CHANGELOG.md
@@ -9,4 +9,4 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- `@fmss/commitlint-plugin` package
+- `@fmss/commitlint-config` package